twisper/requirements.txt
appdirs==1.4.3
asgiref==3.3.1
CacheControl==0.12.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
cycler==0.10.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.4
gunicorn==20.0.4
html5lib==1.0.1
idna==2.10
ipaddr==2.2.0
joblib==1.0.0
kiwisolver==1.3.1
lockfile==0.12.2
matplotlib==3.3.3
msgpack==0.6.2
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
packaging==20.3
pandas==1.2.0
pep517==0.8.2
Pillow==8.1.0
progress==1.5
psycopg2-binary==2.8.6
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytoml==0.1.21
pytz==2020.5
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
tqdm==4.56.0
tweepy==3.10.0
urllib3==1.26.2
webencodings==0.5.1